UniqSupply: make mkSplitUniqSupply thread-safe
unsafeInterleaveIO is used instead of unsafeDupableInterleaveIO because
in a multithreaded setting there is a chance that a mk_supply thunk
could be evaluated by two simultaneous threads.

Also, the UniqSupply counter, GenSymCounter is now incremented
atomically using the RTS's atomic_inc().

To mitigate the extra overhead of unsafeInterleaveIO in the
single-threaded compiler, noDuplicate# is changed to exit early when
n_capabilities == 1. This should be safe, I think.